+++ src/types/date-info.ts
@@ -1,0 +1,7 @@
+import { Post } from "./post"
+
+export interface IDateInfo {
+    day: number
+    date?: number
+    posts?: Post[]
+  }
